+++ Masrofy/Masrofy.BLL/Models/PlanVM.cs
@@ -14,11 +14,11 @@ namespace Masrofy.BLL.Models
     {
      
         public int Id { get; set; } //PK 
-        public double Income { get; set; }
-        public double Charity { get; set; } //حساب صدقة
-        public double SavingAmount { get; set; }
-        public double PersonalAccount { get; set; } //حساب شخصي 
-        public double Obligation { get; set; } //التزامات
+        public decimal Income { get; set; }
+        public decimal Charity { get; set; } //حساب صدقة
+        public decimal SavingAmount { get; set; }
+        public decimal PersonalAccount { get; set; } //حساب شخصي 
+        public decimal Obligation { get; set; } //التزامات
         public DateTime Month { get; set; }
         public string IdentityUserId { get; set; } //FK
